repalce pricing table eight with seven from registry and documentation
