+++ content/docs/components/index.mdx
@@ -15,8 +15,7 @@ Professional pricing table components designed for conversion optimization and c
 | **[Pricing Table Four](/docs/components/pricing-table/pricing-table-four)**   | Modern gradient pricing table with contemporary design and smooth animations |
 | **[Pricing Table Five](/docs/components/pricing-table/pricing-table-five)**   | Modern pricing table with contact us plan                                    |
 | **[Pricing Table Six](/docs/components/pricing-table/pricing-table-six)**   | Contemporary pricing table featuring gradient styling, smooth animations, and an enterprise-focused plan   |
-| **[Pricing Table Seven](/docs/components/pricing-table/pricing-table-seven)**   | A clean and modern pricing plans UI with three tiered cards, a monthly/yearly toggle, and clear call-to-action buttons for different team sizes |
-| **[Pricing Table Eight](/docs/components/pricing-table/pricing-table-eight)**   | Feature-comparison pricing table with aligned cards and user slider for comprehensive plan comparison |
+| **[Pricing Table Seven](/docs/components/pricing-table/pricing-table-seven)**   | Feature-comparison pricing table with aligned cards and user slider for comprehensive plan comparison |
 ### Subscription Management
 
 Comprehensive subscription management interfaces for billing dashboards and user account management.
