+++ content/docs/meta.json
@@ -14,7 +14,6 @@
     "components/pricing-table/pricing-table-five",
     "components/pricing-table/pricing-table-six",
     "components/pricing-table/pricing-table-seven",
-    "components/pricing-table/pricing-table-eight",
     "---Subscription Management---",
     "components/manage-subscription/index",
     "components/manage-subscription/usage-based-pricing",
