+++ public/r/all.json
@@ -12,7 +12,6 @@
     "@billingsdk/pricing-table-five",
     "@billingsdk/pricing-table-six",
     "@billingsdk/pricing-table-seven",
-    "@billingsdk/pricing-table-eight",
     "@billingsdk/cancel-subscription-dialog",
     "@billingsdk/cancel-subscription-card",
     "@billingsdk/update-plan-dialog",
