+++ registry.json
@@ -30,7 +30,6 @@
                 "@billingsdk/pricing-table-five",
                 "@billingsdk/pricing-table-six",
                 "@billingsdk/pricing-table-seven",
-                "@billingsdk/pricing-table-eight",
                 "@billingsdk/cancel-subscription-dialog",
                 "@billingsdk/cancel-subscription-card",
                 "@billingsdk/update-plan-dialog",
@@ -323,38 +322,6 @@
                     "path": "src/registry/billingsdk/demo/pricing-table-seven-demo.tsx",
                     "type": "registry:component",
                     "target": "components/pricing-table-seven-demo.tsx"
-                },
-                {
-                    "path": "src/registry/lib/billingsdk-config.ts",
-                    "type": "registry:lib",
-                    "target": "lib/billingsdk-config.ts"
-                }
-            ],
-            "dependencies": [
-                "lucide-react",
-                "motion"
-            ],
-            "registryDependencies": [
-                "button",
-                "card",
-                "utils"
-            ]
-        },
-        {
-            "name": "pricing-table-eight",
-            "type": "registry:block",
-            "title": "Pricing Table Eight",
-            "description": "Feature-comparison pricing table with aligned cards and user slider",
-            "files": [
-                {
-                    "path": "src/registry/billingsdk/pricing-table-eight.tsx",
-                    "type": "registry:component",
-                    "target": "components/billingsdk/pricing-table-eight.tsx"
-                },
-                {
-                    "path": "src/registry/billingsdk/demo/pricing-table-eight-demo.tsx",
-                    "type": "registry:component",
-                    "target": "components/pricing-table-eight-demo.tsx"
                 },
                 {
                     "path": "src/registry/lib/billingsdk-config.ts",
