--- src/components/billingsdk/pricing-table-eight.tsx
@@ -1,8 +1,0 @@
-"use client";
-export {
-  PricingTableEight,
-  type PricingTableEightProps,
-  type PricingTableEightPlan,
-  type FeatureCategory,
-  type FeatureItemRecord
-} from "@/registry/billingsdk/pricing-table-eight";
