+++ src/components/billingsdk/pricing-table-seven.tsx
@@ -1,3 +1,8 @@
 "use client";
-
-export { PricingTableSeven, type PricingTableSevenProps } from "@/registry/billingsdk/pricing-table-seven"
+export {
+  PricingTableSeven,
+  type PricingTableSevenProps,
+  type PricingTableSevenPlan,
+  type FeatureCategory,
+  type FeatureItemRecord
+} from "@/registry/billingsdk/pricing-table-seven";
