+++ src/mdx-components.tsx
@@ -46,8 +46,8 @@ import { PricingTableSixDemo } from '@/components/pricing-table-six-demo';
 import { UsageBasedPricingDemo } from '@/components/usage-based-pricing-demo';
 import { DetailedUsageTableDemo } from '@/components/detailed-usage-table-demo';
 import { PricingTableSevenDemo } from '@/components/pricing-table-seven-demo';
-import { PricingTableEightDemo } from '@/components/pricing-table-eight-demo';
 import {PaymentCardDemo} from '@/components/payment-card-demo';
+import { PricingTableSevenMinimalDemo } from '@/components/pricing-table-seven-minimal-demo';
 
 
 // use this function to get MDX components, you will need it for rendering MDX
@@ -100,7 +100,7 @@ export function getMDXComponents(components?: MDXComponents): MDXComponents {
 	  PricingTableSixDemo,
     DetailedUsageTableDemo,
     PricingTableSevenDemo,
-    PricingTableEightDemo,
+    PricingTableSevenMinimalDemo,
 	  BillingSettings2Demo,
     UsageBasedPricingDemo
   };
